requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
python-dotenv==1.0.0
schedule==1.2.0
langgraph==0.0.15
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1
python-multipart==0.0.6
requests
flask
redis
sentence-transformers
numpy
chromadb
neo4j
tiktoken
beautifulsoup4
scipy
ollama
